fix(oauth2-mock-seedwork): tests - make jwt.test typesafe and use bracket access
Set kid default to avoid exactOptionalPropertyTypes mismatch and use bracket notation when inspecting parsed JWT payloads.
