+++ packages/cellix/server-oauth2-mock-seedwork/src/utils.ts
@@ -1,0 +1,29 @@
+/** Utility helpers for server-oauth2-mock-seedwork */
+
+export const SAFE_NAME_RE = /^[a-zA-Z0-9_-]+$/;
+
+export const normalizeUrl = (value: string) => {
+	try {
+		const url = new URL(value);
+		const pathname = url.pathname.replace(/\/$/, '') || '/';
+		const params = new URLSearchParams(url.search);
+		params.sort();
+		const search = params.toString() ? `?${params.toString()}` : '';
+		return `${url.origin}${pathname}${search}`;
+	} catch {
+		return value;
+	}
+};
+
+export const normalizeOrigin = (value: string) => {
+	try {
+		const url = new URL(value);
+		return `${url.protocol}//${url.host}`;
+	} catch {
+		return value;
+	}
+};
+
+export function normalizeBaseUrl(url: string): string {
+	return url.replace(/\/$/, '');
+}
